Remove file migration command #5
