+++ tests/TestCase.php
@@ -1,7 +1,5 @@
 <?php
 
-use Illuminate\Http\Request;
-
 class TestCase extends Illuminate\Foundation\Testing\TestCase
 {
     /**
@@ -21,18 +19,6 @@ class TestCase extends Illuminate\Foundation\Testing\TestCase
         $app = require __DIR__.'/../bootstrap/app.php';
 
         $app->make(Illuminate\Contracts\Console\Kernel::class)->bootstrap();
-
-        Route::post('/user', ['middleware' => 'validate:App\Http\Validators\UserValidate',
-            function () {
-                return response()->json(null, 200);
-            }
-        ]);
-
-        Route::post('/comments', ['middleware' => 'camelToSnake',
-            function (Request $request) {
-                return response()->json($request->all(), 200);
-            }
-        ]);
 
         return $app;
     }
